Raw data filtering WIP

+++ toimivuus/toimivuus.py
@@ -1,5 +1,8 @@
+import io
 import logging as log
 import os
+import pandas
+import zstandard
 from azure.storage.blob import BlobServiceClient, ContainerClient
 from enum import Enum, auto
 from datetime import datetime
@@ -45,7 +48,7 @@ class RawHfpFile:
             self.blob_client = container_client.get_blob_client(self.raw_file_name)
         else:
             self.blob_client = None
-        self.data = None
+        self.dataframe: pandas.DataFrame = None
     
     def remote_exists(self) -> bool:
         return self.blob_client is not None and self.blob_client.exists()
@@ -57,9 +60,36 @@ class RawHfpFile:
         if not self.remote_exists():
             log.warning(f'Remote file missing: cannot download {self.raw_file_name}')
             return
+        if self.local_exists():
+            log.info(f'{self.raw_file_name} already exists, skipping download')
+            return
         with open(self.local_path, 'wb') as f:
             f.write(self.blob_client.download_blob().readall())
             log.info(f'{self.local_path} downloaded')
+
+    def read_dataframe(self) -> None:
+        """Read raw zstd-compressed file to pandas DataFrame."""
+        if not self.local_exists():
+            log.warning(f'{self.local_path} missing, cannot read to data frame')
+            return
+        with open(self.local_path, 'rb') as fh:
+            dctx = zstandard.ZstdDecompressor()
+            stream_reader = dctx.stream_reader(fh)
+            text_stream = io.TextIOWrapper(stream_reader, encoding='utf-8')
+            self.dataframe = pandas.read_csv(text_stream)
+            
+    def filter_dataframe(self, routes: List[str] = None, columns: List[str] = None) -> None:
+        """Filter the dataframe, keeping only given columns and rows with given routes.
+        With empty params, all columns / all rows are kept, respectively."""
+        if self.dataframe is None:
+            log.warning('No dataframe read, cannot filter')
+            return
+        orig_shape = self.dataframe.shape
+        if routes is not None:
+            self.dataframe = self.dataframe.loc[self.dataframe['route'].isin(routes)]
+        if columns is not None:
+            self.dataframe = self.dataframe[columns]
+        log.info(f'{self.raw_file_name}: {str(orig_shape)} filtered to {str(self.dataframe.shape)}')
 
     def delete_local(self) -> None:
         if not self.local_exists():
@@ -73,7 +103,7 @@ class RawHfpDump:
     
     def __init__(self, 
                  dump_date_hour: str, 
-                 event_types: List[str], ):
+                 event_types: List[str]):
         dump_datetime: datetime = datetime.strptime(dump_date_hour,
                                                     '%Y-%m-%dT%H')
         self.event_types: List[EventType] = [EventType[evt] for evt in event_types]
@@ -85,8 +115,14 @@ class RawHfpDump:
         """Download compressed csv files by event types to cache."""
         for rhf in self.raw_hfp_files:
             rhf.download_remote()
+            
+    def make_filtered_dataframes(self, routes: List[str], columns: List[str]) -> None:
+        """Read raw data files and filter with columns and routes."""
+        for rhf in self.raw_hfp_files:
+            rhf.read_dataframe()
+            rhf.filter_dataframe(routes=routes, columns=columns)
     
-    def create_merged_file(self, columns: List[str], routes: List[str]) -> None:
+    def create_merged_file(self) -> None:
         """Merge csv files by event type into one file that contains
         the specified columns and rows with specified routes only."""
         pass
